PR feedback - merge email-service-templater.ts and report-emailer.ts
